requirements.txt
torch
torchvision
timm
numpy
pandas
scikit-learn
tqdm
Pillow
matplotlib
imageio
imageio-ffmpeg
PyYAML
